+++ package.json
@@ -25,6 +25,8 @@
         ]
     },
     "dependencies": {
+        "apollo-cache-inmemory": "^1.2.10",
+        "apollo-link-batch-http": "^1.2.3",
         "apollo-server-express": "^2.0.7",
         "express": "^4.16.3",
         "graphql": "^14.0.2"
